feat(arc-412): move admin pages from admin module to next pages
