+++ public/locales/nl/common.json
@@ -17,35 +17,36 @@
   "modules/account/const/my-history___toegang-tot": "Toegang tot",
   "modules/account/const/my-history___toegang-van": "Toegang van",
   "modules/account/layouts/account-layout/account-layout___account": "Account",
-  "modules/admin/reading-rooms/components/admin-navigation/admin-navigation___aanvragen": "Aanvragen",
-  "modules/admin/reading-rooms/components/admin-navigation/admin-navigation___actieve-bezoekers": "Actieve bezoekers",
-  "modules/admin/reading-rooms/components/admin-navigation/admin-navigation___alle-leeszalen": "Alle leeszalen",
-  "modules/admin/reading-rooms/components/admin-navigation/admin-navigation___leeszalenbeheer": "Leeszalenbeheer",
-  "modules/admin/reading-rooms/pages/reading-room-edit/reading-room-edit___instellingen": "Instellingen",
-  "modules/admin/reading-rooms/pages/reading-rooms-overview/reading-rooms-overview___activeren": "Activeren",
-  "modules/admin/reading-rooms/pages/reading-rooms-overview/reading-rooms-overview___alle-leeszalen": "Alle leeszalen",
-  "modules/admin/reading-rooms/pages/reading-rooms-overview/reading-rooms-overview___deactiveren": "Deactiveren",
-  "modules/admin/reading-rooms/pages/reading-rooms-overview/reading-rooms-overview___emailadres": "Emailadres",
-  "modules/admin/reading-rooms/pages/reading-rooms-overview/reading-rooms-overview___geactiveerd-op": "Geactiveerd op",
+  "modules/admin/const/requests___alle": "alle",
+  "modules/admin/const/requests___emailadres": "emailadres",
+  "modules/admin/const/requests___geweigerd": "geweigerd",
+  "modules/admin/const/requests___goedgekeurd": "goedgekeurd",
+  "modules/admin/const/requests___leeszaal": "leeszaal",
+  "modules/admin/const/requests___naam": "naam",
+  "modules/admin/const/requests___open": "open",
+  "modules/admin/const/requests___status": "status",
+  "modules/admin/const/requests___tijdstip": "tijdstip",
+  "modules/admin/const/routing___aanvragen": "aanvragen",
+  "modules/admin/const/routing___actieve-bezoekers": "actieve-bezoekers",
+  "modules/admin/const/routing___alle-leeszalen": "alle-leeszalen",
+  "modules/admin/const/routing___leeszalenbeheer": "leeszalenbeheer",
+  "modules/admin/const/spaces___activeren": "activeren",
+  "modules/admin/const/spaces___deactiveren": "deactiveren",
+  "modules/admin/const/spaces___emailadres": "emailadres",
+  "modules/admin/const/spaces___geactiveerd-op": "geactiveerd-op",
+  "modules/admin/const/spaces___leeszaal": "leeszaal",
+  "modules/admin/const/spaces___publicatiestatus": "publicatiestatus",
+  "modules/admin/const/spaces___telefoonnummer": "telefoonnummer",
+  "modules/admin/const/spaces___terug-naar-in-aanvraag": "terug-naar-in-aanvraag",
+  "modules/admin/const/visitors___goedgekeurd-door": "goedgekeurd-door",
+  "modules/admin/const/visitors___leeszaal": "leeszaal",
+  "modules/admin/const/visitors___naam": "naam",
+  "modules/admin/const/visitors___toegang": "toegang",
+  "modules/admin/layouts/admin-layout/admin-layout___admin": "Admin",
   "modules/admin/reading-rooms/pages/reading-rooms-overview/reading-rooms-overview___geen-leeszalen-gevonden": "Geen leeszalen gevonden",
-  "modules/admin/reading-rooms/pages/reading-rooms-overview/reading-rooms-overview___leeszaal": "Leeszaal",
-  "modules/admin/reading-rooms/pages/reading-rooms-overview/reading-rooms-overview___publicatiestatus": "Publicatiestatus",
-  "modules/admin/reading-rooms/pages/reading-rooms-overview/reading-rooms-overview___telefoonnummer": "Telefoonnummer",
-  "modules/admin/reading-rooms/pages/reading-rooms-overview/reading-rooms-overview___terug-naar-in-aanvraag": "Terug naar \"in aanvraag\"",
   "modules/admin/reading-rooms/pages/reading-rooms-overview/reading-rooms-overview___zoek": "Zoek",
-  "modules/admin/reading-rooms/pages/requests/requests___aanvragen": "Aanvragen",
-  "modules/admin/reading-rooms/pages/requests/requests___emailadres": "Emailadres",
-  "modules/admin/reading-rooms/pages/requests/requests___leeszaal": "Leeszaal",
-  "modules/admin/reading-rooms/pages/requests/requests___naam": "Naam",
-  "modules/admin/reading-rooms/pages/requests/requests___status": "Status",
-  "modules/admin/reading-rooms/pages/requests/requests___tijdstip": "Tijdstip",
-  "modules/admin/reading-rooms/pages/visitors/visitors___actieve-bezoekers": "Actieve bezoekers",
   "modules/admin/reading-rooms/pages/visitors/visitors___er-zijn-geen-actieve-bezoekers": "Er zijn geen actieve bezoekers.",
-  "modules/admin/reading-rooms/pages/visitors/visitors___goedgekeurd-door": "Goedgekeurd door",
   "modules/admin/reading-rooms/pages/visitors/visitors___laden": "Laden",
-  "modules/admin/reading-rooms/pages/visitors/visitors___leeszaal": "Leeszaal",
-  "modules/admin/reading-rooms/pages/visitors/visitors___naam": "Naam",
-  "modules/admin/reading-rooms/pages/visitors/visitors___toegang": "Toegang",
   "modules/admin/reading-rooms/pages/visitors/visitors___zoek": "Zoek",
   "modules/auth/components/auth-modal/auth-modal___inloggen-met-het-archief-account": "Inloggen met Het Archief-account",
   "modules/auth/components/auth-modal/auth-modal___inloggen-of-registreren": "Inloggen of registreren",
@@ -327,6 +328,19 @@
   "pages/account/mijn-profiel/index___mijn-profiel-meta-omschrijving": "Mijn profiel meta omschrijving",
   "pages/account/mijn-profiel/index___voornaam": "Voornaam",
   "pages/account/mijn-profiel/index___wijzig-mijn-gegevens": "Wijzig mijn gegevens",
+  "pages/admin/leeszalenbeheer/aanvragen/index___actieve": "Actieve",
+  "pages/admin/leeszalenbeheer/aanvragen/index___actieve-bezoekers": "Actieve bezoekers",
+  "pages/admin/leeszalenbeheer/aanvragen/index___actieve-meta-omscrhijving": "Actieve meta omscrhijving",
+  "pages/admin/leeszalenbeheer/bezoekers/index___alle-leeszalen": "Alle leeszalen",
+  "pages/admin/leeszalenbeheer/bezoekers/index___alle-leeszalen-meta-omscrhijving": "Alle leeszalen meta omscrhijving",
+  "pages/admin/leeszalenbeheer/leeszalen/index___alle-leeszalen": "Alle leeszalen",
+  "pages/admin/leeszalenbeheer/leeszalen/index___alle-leeszalen-meta-omscrhijving": "Alle leeszalen meta omscrhijving",
+  "pages/admin/leeszalenbeheer/leeszalen/index___de-status-werd-succesvol-aangepast": "de status werd succesvol aangepast",
+  "pages/admin/leeszalenbeheer/leeszalen/index___er-ging-iets-mis": "er ging iets mis",
+  "pages/admin/leeszalenbeheer/leeszalen/index___er-is-een-fout-opgetreden-tijdens-het-aanpassen-van-de-status": "er is een fout opgetreden tijdens het aanpassen van de status",
+  "pages/admin/leeszalenbeheer/leeszalen/index___succes": "succes",
+  "pages/admin/leeszalenbeheer/leeszalen/slug/index___instellingen": "Instellingen",
+  "pages/admin/leeszalenbeheer/leeszalen/slug/index___instellingen-meta-omscrhijving": "Instellingen meta omscrhijving",
   "pages/beheer/aanvragen/index___aanvragen": "Aanvragen",
   "pages/beheer/aanvragen/index___aanvragen-meta-omschrijving": "Aanvragen meta omschrijving",
   "pages/beheer/aanvragen/index___er-zijn-geen-geannuleerde-aanvragen": "Er zijn geen geannuleerde aanvragen.",
