+++ src/modules/admin/layouts/AdminLayout/AdminLayout.types.ts
@@ -1,0 +1,5 @@
+import { DefaultComponentProps } from '@shared/types';
+
+export interface AdminLayoutProps extends DefaultComponentProps {
+	contentTitle?: string;
+}
