+++ src/modules/admin/layouts/AdminLayout/index.ts
@@ -1,0 +1,2 @@
+export { default as AdminLayout } from './AdminLayout';
+export * from './AdminLayout.types';
