--- src/modules/admin/navigation/index.tsx
@@ -1,6 +1,0 @@
-import { NAVIGATION_PATHS } from './const';
-import { NavigationOverview } from './pages';
-
-export const NAVIGATION_ROUTES = [
-	{ path: NAVIGATION_PATHS.overview, component: NavigationOverview },
-];
