--- src/modules/admin/navigation/pages/NavigationOverview/index.ts
@@ -1,1 +1,0 @@
-export { default as NavigationOverview } from './NavigationOverview';
