--- src/modules/admin/navigation/pages/index.ts
@@ -1,1 +1,0 @@
-export * from './NavigationOverview';
