--- src/modules/admin/readingRooms/components/AdminNavigation/AdminNavigation.types.ts
@@ -1,3 +1,0 @@
-import { DefaultComponentProps } from '@shared/types';
-
-export type AdminNavigationProps = DefaultComponentProps;
