--- src/modules/admin/readingRooms/components/AdminNavigation/index.ts
@@ -1,2 +1,0 @@
-export { default as AdminNavigation } from './AdminNavigation';
-export * from './AdminNavigation.types';
