--- src/modules/admin/readingRooms/components/index.ts
@@ -1,1 +1,0 @@
-export * from './AdminNavigation';
