--- src/modules/admin/readingRooms/const/index.ts
@@ -1,11 +1,0 @@
-const ROOT = 'leeszalenbeheer';
-const DETAIL = ':slug';
-
-export const READING_ROOMS_PATHS = {
-	overview: `/${ROOT}`,
-	leeszalen: `/${ROOT}/leeszalen`,
-	aanvragen: `/${ROOT}/aanvragen`,
-	bezoekers: `/${ROOT}/bezoekers`,
-	create: `/${ROOT}/${DETAIL}/aanmaken`,
-	edit: `/${ROOT}/${DETAIL}/bewerken`,
-};
