--- src/modules/admin/readingRooms/pages/ReadingRoomEdit/index.ts
@@ -1,1 +1,0 @@
-export { default as ReadingRoomEdit } from './ReadingRoomEdit';
