--- src/modules/admin/readingRooms/pages/ReadingRoomsOverview/index.ts
@@ -1,2 +1,0 @@
-export { default as ReadingRoomsOverview } from './ReadingRoomsOverview';
-export * from './ReadingRoomsOverview.const';
