--- src/modules/admin/readingRooms/pages/Requests/index.ts
@@ -1,2 +1,0 @@
-export { default as Requests } from './Requests';
-export * from './Requests.const';
