--- src/modules/admin/readingRooms/pages/Visitors/index.ts
@@ -1,2 +1,0 @@
-export { default as Visitors } from './Visitors';
-export * from './Visitors.const';
