--- src/modules/admin/readingRooms/pages/index.ts
@@ -1,4 +1,0 @@
-export * from './ReadingRoomEdit';
-export * from './ReadingRoomsOverview';
-export * from './Requests';
-export * from './Visitors';
