+++ src/modules/navigation/components/Navigation/Navigation.consts.tsx
@@ -178,7 +178,7 @@ export const getNavigationItemsLeft = (
 		],
 	},
 	{
-		node: renderLink('Admin', '/admin/leeszalenbeheer/leeszalen', {
+		node: renderLink('Admin', '', {
 			className: linkCls([
 				'u-color-black',
 				'u-color-white:md',
@@ -187,5 +187,25 @@ export const getNavigationItemsLeft = (
 		}),
 		id: 'nav__admin',
 		active: currentPath.startsWith('/admin'),
+		children: [
+			{
+				node: renderLink('Alle leeszalen', '/admin/leeszalenbeheer/leeszalen', {
+					className: dropdownCls(),
+				}),
+				id: 'nav__admin--leeszalen',
+			},
+			{
+				node: renderLink('Aanvragen', '/admin/leeszalenbeheer/aanvragen', {
+					className: dropdownCls(),
+				}),
+				id: 'nav__admin--aanvragen',
+			},
+			{
+				node: renderLink('Actieve bezoekers', '/admin/leeszalenbeheer/bezoekers', {
+					className: dropdownCls(),
+				}),
+				id: 'nav__admin--bezoekers',
+			},
+		],
 	},
 ];
