+++ src/modules/shared/const/routes.ts
@@ -3,4 +3,5 @@ export const ROUTES = {
 	termsOfService: '/gebruiksvoorwaarden',
 	myCollections: '/account/mijn-mappen',
 	visitRequested: '/:slug/toegang-aangevraagd',
+	adminEditSpace: 'admin/leeszalenbeheer/leeszalen/:slug',
 };
