+++ tsconfig.json
@@ -5,6 +5,9 @@
       "@account/*": [
         "modules/account/*"
       ],
+      "@admin/*": [
+        "modules/admin/*"
+      ],
       "@auth/*": [
         "modules/auth/*"
       ],
